test: move bail tests to test/cli
