--- test/bail/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@vitest/test-bail",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "test": "vitest"
-  },
-  "devDependencies": {
-    "@vitest/browser": "workspace:*",
-    "vite": "latest",
-    "vitest": "workspace:*",
-    "webdriverio": "latest"
-  }
-}
