+++ test/cli/fixtures/bail/vitest.config.ts
